Extrair codigo de teste de objects para arquivo separado.
