+++ src/br/com/alura/bytebank/main.kt
@@ -1,30 +1,7 @@
 package br.com.alura.bytebank
 
-import br.com.alura.bytebank.modelo.*
-import br.com.alura.bytebank.teste.testaContasDiferentes
+import br.com.alura.bytebank.teste.testaObjects
 
 fun main() {
-
-    //link doc "Object Expressions and Declarations": https://kotlinlang.org/docs/reference/object-declarations.html
-    //Object expressions
-    val fran = object: Autenticavel {
-        val nome: String = "Fran"
-        val cpf: String = "111.111.111-11"
-        val senha: Int = 1000
-
-        override fun autentica(senha: Int) = this.senha == senha
-    }
-
-    val sistemaInterno = SistemaInterno()
-    sistemaInterno.entra(fran, 1000)
-
-    println("nome do cliente ${fran.nome}")
-
-    val alex = Cliente(nome = "Alex", cpf = "", senha = 1)
-    val contaPoupanca = ContaPoupanca(titular = alex, numero = 1000)
-    val contaCorrente = ContaCorrente(titular = alex, numero = 1001)
-
-    testaContasDiferentes()
-
-    println("Total de contas: ${Conta.total}")
+    testaObjects()
 }
